server/requirements.txt
asgi-redis==1.3.0
django_channels==0.7.0
django_debug_toolbar==1.7
Django==1.11
django_environ==0.4.1
redis==2.10.5
channels==1.1.3
environ==1.0
django_model_utils==3.0.0
djangorestframework==3.6.2
tox==2.7.0
flake8==3.3.0
pytest==3.0.7
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-mock==1.6.0
pytest-redis==1.2.1
pytest-testmon==0.9.4
pytest-watch==4.1.0
psycopg2==2.7.1
django-allauth==0.31.0
django-channels-panel==0.0.5
